requirements.txt
requests==2.32.3
numpy==1.26.4
pandas
matplotlib
tqdm
fastparquet
gdown
mlflow==3.8.0
python-dotenv==1.0.1
pyyaml==6.0.1
optuna
transformers
